+++ source/classes/constants.py
@@ -114,6 +114,8 @@ SETTINGSTOPROCESS = {
       "uwpalettes": "uw_palettes",
       "reduce_flashing": "reduce_flashing",
       "shuffle_sfx": "shuffle_sfx",
+      'msu_resume': 'msu_resume',
+      'collection_rate': 'collection_rate',
     },
     "generation": {
       "bps": "bps",